Switch to using a singular BrightnessMeter due to poor Linux performance.
