+++ TempoAgents/Source/TempoAgentsShared/Private/TempoAgentsWorldSubsystem.cpp
@@ -5,7 +5,11 @@
 #include "EngineUtils.h"
 #include "MassTrafficLightRegistrySubsystem.h"
 #include "TempoAgentsShared.h"
+#include "TempoBrightnessMeter.h"
+#include "TempoBrightnessMeterComponent.h"
 #include "TempoIntersectionInterface.h"
+#include "TempoRoadInterface.h"
+#include "Kismet/GameplayStatics.h"
 
 void UTempoAgentsWorldSubsystem::SetupTrafficControllers()
 {
@@ -38,6 +42,70 @@ void UTempoAgentsWorldSubsystem::SetupTrafficControllers()
 	}
 }
 
+void UTempoAgentsWorldSubsystem::SetupBrightnessMeter()
+{
+	UWorld& World = GetWorldRef();
+
+	// If there's already a TempoBrightnessMeter in the world, we'll just use that.
+	const AActor* FoundBrightnessMeterActor = UGameplayStatics::GetActorOfClass(&World, ATempoBrightnessMeter::StaticClass());
+	if (FoundBrightnessMeterActor != nullptr)
+	{
+		return;
+	}
+
+	TArray<AActor*> RoadActors;
+	for (AActor* Actor : TActorRange<AActor>(&World))
+	{
+		if (!IsValid(Actor))
+		{
+			continue;
+		}
+
+		if (!Actor->Implements<UTempoRoadInterface>())
+		{
+			continue;
+		}
+
+		RoadActors.Add(Actor);
+	}
+
+	if (RoadActors.IsEmpty())
+	{
+		UE_LOG(LogTempoAgentsShared, Log, TEXT("No RoadActors found to be used as a spawn anchor in UTempoAgentsWorldSubsystem::SetupBrightnessMeter."));
+		return;
+	}
+
+	// Sort the road actors so that we can make a deterministic (given the state of the roads in the world) selection
+	// to which we will anchor the BrightnessMeter.
+	RoadActors.Sort([](const AActor& Actor1, const AActor& Actor2)
+	{
+		return Actor1.GetName() < Actor2.GetName();
+	});
+
+	AActor& AnchorRoadActor = *RoadActors[0];
+
+	const int32 RoadStartEntranceLocationControlPointIndex = ITempoRoadInterface::Execute_GetTempoStartEntranceLocationControlPointIndex(&AnchorRoadActor);
+	
+	const FVector RoadStartLocation = ITempoRoadInterface::Execute_GetTempoControlPointLocation(&AnchorRoadActor, RoadStartEntranceLocationControlPointIndex, ETempoCoordinateSpace::World);
+	const FVector RoadForwardVector = ITempoRoadInterface::Execute_GetTempoControlPointTangent(&AnchorRoadActor, RoadStartEntranceLocationControlPointIndex, ETempoCoordinateSpace::World).GetSafeNormal();
+	const FVector RoadRightVector = ITempoRoadInterface::Execute_GetTempoControlPointRightVector(&AnchorRoadActor, RoadStartEntranceLocationControlPointIndex, ETempoCoordinateSpace::World);
+	const FVector RoadUpVector = ITempoRoadInterface::Execute_GetTempoControlPointUpVector(&AnchorRoadActor, RoadStartEntranceLocationControlPointIndex, ETempoCoordinateSpace::World);
+
+	const float RoadWidth = ITempoRoadInterface::Execute_GetTempoRoadWidth(&AnchorRoadActor);
+	const float TotalLateralOffset = RoadWidth * 0.5f + BrightnessMeterLateralOffset;
+	
+	const FVector SpawnLocation = RoadStartLocation + RoadForwardVector * BrightnessMeterLongitudinalOffset + RoadRightVector * TotalLateralOffset + RoadUpVector * BrightnessMeterVerticalOffset;
+
+	// Face the BrightnessMeter back towards the road.
+	const FRotator SpawnRotation = FMatrix(-RoadRightVector, RoadForwardVector, RoadUpVector, FVector::ZeroVector).Rotator();
+
+	FActorSpawnParameters SpawnParams;
+	SpawnParams.SpawnCollisionHandlingOverride = ESpawnActorCollisionHandlingMethod::AlwaysSpawn;
+
+	ATempoBrightnessMeter* BrightnessMeterActor = GetWorld()->SpawnActor<ATempoBrightnessMeter>(ATempoBrightnessMeter::StaticClass(), SpawnLocation, SpawnRotation, SpawnParams);
+	ensureMsgf(BrightnessMeterActor != nullptr, TEXT("Must spawn valid BrightnessMeterActor in UTempoAgentsWorldSubsystem::SetupBrightnessMeter."));
+}
+
 void UTempoAgentsWorldSubsystem::OnWorldBeginPlay(UWorld& InWorld)
 {
 	Super::OnWorldBeginPlay(InWorld);
@@ -47,4 +115,6 @@ void UTempoAgentsWorldSubsystem::OnWorldBeginPlay(UWorld& InWorld)
 	// For reference, UWorldSubsystem::OnWorldBeginPlay is called before BeginPlay is called on all the Actors
 	// in the level, and UWorld::OnWorldBeginPlay is called after BeginPlay is called on all the Actors in the level.
 	GetWorld()->OnWorldBeginPlay.AddUObject(this, &UTempoAgentsWorldSubsystem::SetupTrafficControllers);
+
+	GetWorld()->OnWorldBeginPlay.AddUObject(this, &UTempoAgentsWorldSubsystem::SetupBrightnessMeter);
 }
